Fix quoting of empty string in `to nuon`
Closes #7631

The fix is absolutely trivial but the whole thing is a perfect lecture
about things that can go wrong in the initial test design.

The proptest tests never generate the whole string but only characters
in a larger string or a single character.
The main catch all test `quotes_some_strings_necessarily` is not
powerful enough to catch the parser quirk that `[,]` is equivalent to
an empty list and the comma or the empty string get ignored.
This only matters in a table as seen in #7631.