Name fields on `ShellError::OperatorOverflow`
Instead of using tuple struct variants go for named struct variants to avoid
confusion when creating `ShellError`s.

The goal is to better document the errors and make them both more useful
to the implementers as well as the users.
A future goal of this effort is to enable serialization of errors to
nushell records for finer-grained error handling.