Comment about issues with `TypeMismatch*`
The two variants are redundant. Naming the fields on `TypeMismatch` would
be effectively the same as `TypeMismatchGenericMessage`

We should make `TypeMismatch` more helpful anyways. (e.g. including both
types and make sure spans are useful)

When doing the joining it is worth checking for places where it is not a
type error but something else.