[experiment] Disable all property-based tests
Primarily to study the impact on the coverage I disabled the
property-based tests.
Touches `quickcheck` and `proptest` tests (do the same thing slightly
differently under the hood.)
For now just commented out. If we can stop the flukes and keep the
general coverage we would want to move the proptests into a separate
scheduled fuzz run.